// src/types/queryResponses.ts
export interface Erc721transfer {
    contract: {
        id: string
        _typename: string
    }
    id: string
    timestamp: BigInt
    _typename: string
}

export interface Erc1155transfer {
    contract: {
        id: string
        _typename: string
    }
    id: string
    timestamp: BigInt
    _typename: string
}